+++ forge.config.ts
@@ -40,12 +40,12 @@ const config: ForgeConfig = {
       build: [
         {
           // `entry` is just an alias for `build.lib.entry` in the corresponding file of `config`.
-          entry: 'src/main.ts',
+          entry: 'code/src/main.ts',
           config: 'vite.main.config.ts',
           target: 'main',
         },
         {
-          entry: 'src/preload.ts',
+          entry: 'code/src/preload.ts',
           config: 'vite.preload.config.ts',
           target: 'preload',
         },